requirements.txt
# ── Core scientific stack ─────────────────────────────────────────────────────
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
scikit-learn>=1.4.0

# ── Survival / causal modelling ───────────────────────────────────────────────
lifelines>=0.27.8           # Survival analysis (Weibull AFT, KM, Cox PH)
networkx>=3.3               # SCM DAG definitions and do-calculus traversal
statsmodels>=0.14.0         # GLM, mixed effects, autocorrelation models

# ── Distributed compute ───────────────────────────────────────────────────────
ray[default]>=2.10.0        # Distributed simulation workers + cluster mgmt
# dask[distributed]>=2024.1.0  # Uncomment as alternative to Ray

# ── Experiment tracking ───────────────────────────────────────────────────────
mlflow>=2.12.0              # Experiment tracking, config + artifact versioning

# ── LLM agent layer ───────────────────────────────────────────────────────────
anthropic>=0.25.0           # Anthropic SDK for claude-sonnet-4-6 agents

# ── API serving layer ─────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# ── Config + serialisation ────────────────────────────────────────────────────
pyyaml>=6.0.1
python-dotenv>=1.0.1

# ── Statistical drift detection ───────────────────────────────────────────────
scipy>=1.12.0               # KS-test, already listed; aliased for clarity

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.1.0
pytest-asyncio>=0.23.0
httpx>=0.27.0               # FastAPI test client
pytest-cov>=5.0.0
